requirements.txt
# requirements
jupyter==1.0.0
numpy==1.19.3 # 1.19.5
seaborn==0.11.0
torch==1.8.1 # 1.10.2
gpytorch==1.5.1
pandas==1.4.1
tqdm==4.54.1
jsonargparse==3.3.2
scikit-learn==0.24.2
scikit-optimize==0.8.1
pyro-ppl==1.6.0